Launcher/requirements.txt
PyYAML==6.0.1
kafka-python==2.0.2
typer==0.9.0
docker==7.0.0